requirements.txt
asgiref==3.8.1
boto3==1.34.131
celery==5.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
Django==4.2.15
djangorestframework==3.15.1
pgvector==0.2.5
psycopg[binary]==3.1.18
redis==5.0.4
django-storages==1.14.3
sentence-transformers==2.7.0
torch==2.3.0
transformers==4.41.1
requests==2.32.3
pytest==8.2.2
pytest-django==4.8.0
python-dotenv==1.0.1
uvicorn==0.30.1
